+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### unreleased
+### Fixed
+* Fixed an issue where the chosen output format would not be used when exporting.
+
 ### 0.2.1 - 2023-08-25
 ### Fixed
 * Fixed an issue that prevented exporting with mipmaps forced to "Disabled".